Refactor Contexts to match C# coding guidelines
