+++ Backend/Startup.cs
@@ -1,5 +1,4 @@
-﻿using BackendFramework.Context;
-using BackendFramework.Interfaces;
+﻿using BackendFramework.Interfaces;
 using BackendFramework.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
@@ -13,6 +12,7 @@ using Microsoft.IdentityModel.Tokens;
 using SIL.Lift.Parsing;
 using System;
 using System.Text;
+using BackendFramework.Contexts;
 
 namespace BackendFramework
 {
